requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
ecdsa==0.19.0
Flask==3.1.0
flask-cors==5.0.1
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-sqlalchemy==1.4.1
packaging==24.2
psycopg==3.2.6
pyasn1==0.4.8
PyJWT==2.10.1
python-jose==3.4.0
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
tzdata==2025.1
Werkzeug==3.1.3
